+++ PDF/002.pdf
@@ -1,0 +1,353 @@
+%PDF-1.3
+%·¾­ª
+1 0 obj
+<<
+/Type /Catalog
+/Pages 2 0 R
+>>
+endobj
+2 0 obj
+<<
+/Type /Pages
+/Kids [ 4 0 R ]
+/Count 1
+>>
+endobj
+3 0 obj
+<<
+/Producer (Haru Free PDF Library 2.4.0dev)
+>>
+endobj
+4 0 obj
+<<
+/Type /Page
+/MediaBox [ 0 0 595 841 ]
+/Contents 5 0 R
+/Resources <<
+/ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+/Font <<
+/F1 7 0 R
+>>
+>>
+/Parent 2 0 R
+>>
+endobj
+5 0 obj
+<<
+/Length 6 0 R
+>>
+stream
+/F1 14 Tf
+BT
+416 775 Td
+(Datum:  15.8.2022) Tj
+ET
+50 750 m
+545 750 l
+S
+BT
+59 691 Td
+(337 + 589 =) Tj
+ET
+131 691 m
+191 691 l
+S
+BT
+59 641 Td
+(935 * 172 =) Tj
+ET
+131 641 m
+191 641 l
+S
+BT
+59 591 Td
+(508 + 277 =) Tj
+ET
+131 591 m
+191 591 l
+S
+BT
+59 541 Td
+(840 * 679 =) Tj
+ET
+131 541 m
+191 541 l
+S
+BT
+59 491 Td
+(586 + 304 =) Tj
+ET
+131 491 m
+191 491 l
+S
+BT
+59 441 Td
+(519 \057 25 =) Tj
+ET
+125 441 m
+185 441 l
+S
+BT
+59 391 Td
+(877 \057 403 =) Tj
+ET
+131 391 m
+191 391 l
+S
+BT
+59 341 Td
+(636 * 836 =) Tj
+ET
+131 341 m
+191 341 l
+S
+BT
+59 291 Td
+(153 \057 862 =) Tj
+ET
+131 291 m
+191 291 l
+S
+BT
+59 241 Td
+(727 + 652 =) Tj
+ET
+131 241 m
+191 241 l
+S
+BT
+59 191 Td
+(971 \057 341 =) Tj
+ET
+131 191 m
+191 191 l
+S
+BT
+59 141 Td
+(661 + 276 =) Tj
+ET
+131 141 m
+191 141 l
+S
+BT
+59 91 Td
+(341 - 136 =) Tj
+ET
+131 91 m
+191 91 l
+S
+BT
+238 691 Td
+(83 * 976 =) Tj
+ET
+304 691 m
+364 691 l
+S
+BT
+238 641 Td
+(435 * 915 =) Tj
+ET
+310 641 m
+370 641 l
+S
+BT
+238 591 Td
+(394 * 434 =) Tj
+ET
+310 591 m
+370 591 l
+S
+BT
+238 541 Td
+(414 + 663 =) Tj
+ET
+310 541 m
+370 541 l
+S
+BT
+238 491 Td
+(296 \057 651 =) Tj
+ET
+310 491 m
+370 491 l
+S
+BT
+238 441 Td
+(931 \057 804 =) Tj
+ET
+310 441 m
+370 441 l
+S
+BT
+238 391 Td
+(215 * 883 =) Tj
+ET
+310 391 m
+370 391 l
+S
+BT
+238 341 Td
+(194 * 854 =) Tj
+ET
+310 341 m
+370 341 l
+S
+BT
+238 291 Td
+(115 + 516 =) Tj
+ET
+310 291 m
+370 291 l
+S
+BT
+238 241 Td
+(208 + 209 =) Tj
+ET
+310 241 m
+370 241 l
+S
+BT
+238 191 Td
+(930 - 293 =) Tj
+ET
+310 191 m
+370 191 l
+S
+BT
+238 141 Td
+(684 - 728 =) Tj
+ET
+310 141 m
+370 141 l
+S
+BT
+238 91 Td
+(743 + 475 =) Tj
+ET
+310 91 m
+370 91 l
+S
+BT
+416 691 Td
+(179 * 241 =) Tj
+ET
+488 691 m
+548 691 l
+S
+BT
+416 641 Td
+(370 - 537 =) Tj
+ET
+488 641 m
+548 641 l
+S
+BT
+416 591 Td
+(397 + 820 =) Tj
+ET
+488 591 m
+548 591 l
+S
+BT
+416 541 Td
+(287 + 36 =) Tj
+ET
+482 541 m
+542 541 l
+S
+BT
+416 491 Td
+(182 \057 582 =) Tj
+ET
+488 491 m
+548 491 l
+S
+BT
+416 441 Td
+(418 * 49 =) Tj
+ET
+482 441 m
+542 441 l
+S
+BT
+416 391 Td
+(282 * 609 =) Tj
+ET
+488 391 m
+548 391 l
+S
+BT
+416 341 Td
+(635 \057 540 =) Tj
+ET
+488 341 m
+548 341 l
+S
+BT
+416 291 Td
+(965 + 576 =) Tj
+ET
+488 291 m
+548 291 l
+S
+BT
+416 241 Td
+(209 - 319 =) Tj
+ET
+488 241 m
+548 241 l
+S
+BT
+416 191 Td
+(592 + 850 =) Tj
+ET
+488 191 m
+548 191 l
+S
+BT
+416 141 Td
+(637 - 220 =) Tj
+ET
+488 141 m
+548 141 l
+S
+BT
+416 91 Td
+(334 * 618 =) Tj
+ET
+488 91 m
+548 91 l
+S
+
+endstream
+endobj
+6 0 obj
+2253
+endobj
+7 0 obj
+<<
+/Type /Font
+/BaseFont /Helvetica
+/Subtype /Type1
+/Encoding /StandardEncoding
+>>
+endobj
+xref
+0 8
+0000000000 65535 f
+0000000015 00000 n
+0000000064 00000 n
+0000000123 00000 n
+0000000187 00000 n
+0000000363 00000 n
+0000002670 00000 n
+0000002690 00000 n
+trailer
+<<
+/Root 1 0 R
+/Info 3 0 R
+/Size 8
+>>
+startxref
+2788
+%%EOF
