+++ PDF/002_Lösungen.pdf
@@ -1,0 +1,236 @@
+%PDF-1.3
+%·¾­ª
+1 0 obj
+<<
+/Type /Catalog
+/Pages 2 0 R
+>>
+endobj
+2 0 obj
+<<
+/Type /Pages
+/Kids [ 4 0 R ]
+/Count 1
+>>
+endobj
+3 0 obj
+<<
+/Producer (Haru Free PDF Library 2.4.0dev)
+>>
+endobj
+4 0 obj
+<<
+/Type /Page
+/MediaBox [ 0 0 595 841 ]
+/Contents 5 0 R
+/Resources <<
+/ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+/Font <<
+/F1 7 0 R
+>>
+>>
+/Parent 2 0 R
+>>
+endobj
+5 0 obj
+<<
+/Length 6 0 R
+>>
+stream
+/F1 14 Tf
+BT
+416 775 Td
+(Datum:  15.8.2022) Tj
+ET
+50 750 m
+545 750 l
+S
+BT
+59 691 Td
+(337 + 589 = 926.00) Tj
+ET
+BT
+59 641 Td
+(935 * 172 = 160820.00) Tj
+ET
+BT
+59 591 Td
+(508 + 277 = 785.00) Tj
+ET
+BT
+59 541 Td
+(840 * 679 = 570360.00) Tj
+ET
+BT
+59 491 Td
+(586 + 304 = 890.00) Tj
+ET
+BT
+59 441 Td
+(519 \057 25 = 20.76) Tj
+ET
+BT
+59 391 Td
+(877 \057 403 = 2.18) Tj
+ET
+BT
+59 341 Td
+(636 * 836 = 531696.00) Tj
+ET
+BT
+59 291 Td
+(153 \057 862 = 0.18) Tj
+ET
+BT
+59 241 Td
+(727 + 652 = 1379.00) Tj
+ET
+BT
+59 191 Td
+(971 \057 341 = 2.85) Tj
+ET
+BT
+59 141 Td
+(661 + 276 = 937.00) Tj
+ET
+BT
+59 91 Td
+(341 - 136 = 205.00) Tj
+ET
+BT
+238 691 Td
+(83 * 976 = 81008.00) Tj
+ET
+BT
+238 641 Td
+(435 * 915 = 398025.00) Tj
+ET
+BT
+238 591 Td
+(394 * 434 = 170996.00) Tj
+ET
+BT
+238 541 Td
+(414 + 663 = 1077.00) Tj
+ET
+BT
+238 491 Td
+(296 \057 651 = 0.45) Tj
+ET
+BT
+238 441 Td
+(931 \057 804 = 1.16) Tj
+ET
+BT
+238 391 Td
+(215 * 883 = 189845.00) Tj
+ET
+BT
+238 341 Td
+(194 * 854 = 165676.00) Tj
+ET
+BT
+238 291 Td
+(115 + 516 = 631.00) Tj
+ET
+BT
+238 241 Td
+(208 + 209 = 417.00) Tj
+ET
+BT
+238 191 Td
+(930 - 293 = 637.00) Tj
+ET
+BT
+238 141 Td
+(684 - 728 = -44.00) Tj
+ET
+BT
+238 91 Td
+(743 + 475 = 1218.00) Tj
+ET
+BT
+416 691 Td
+(179 * 241 = 43139.00) Tj
+ET
+BT
+416 641 Td
+(370 - 537 = -167.00) Tj
+ET
+BT
+416 591 Td
+(397 + 820 = 1217.00) Tj
+ET
+BT
+416 541 Td
+(287 + 36 = 323.00) Tj
+ET
+BT
+416 491 Td
+(182 \057 582 = 0.31) Tj
+ET
+BT
+416 441 Td
+(418 * 49 = 20482.00) Tj
+ET
+BT
+416 391 Td
+(282 * 609 = 171738.00) Tj
+ET
+BT
+416 341 Td
+(635 \057 540 = 1.18) Tj
+ET
+BT
+416 291 Td
+(965 + 576 = 1541.00) Tj
+ET
+BT
+416 241 Td
+(209 - 319 = -110.00) Tj
+ET
+BT
+416 191 Td
+(592 + 850 = 1442.00) Tj
+ET
+BT
+416 141 Td
+(637 - 220 = 417.00) Tj
+ET
+BT
+416 91 Td
+(334 * 618 = 206412.00) Tj
+ET
+
+endstream
+endobj
+6 0 obj
+1700
+endobj
+7 0 obj
+<<
+/Type /Font
+/BaseFont /Helvetica
+/Subtype /Type1
+/Encoding /StandardEncoding
+>>
+endobj
+xref
+0 8
+0000000000 65535 f
+0000000015 00000 n
+0000000064 00000 n
+0000000123 00000 n
+0000000187 00000 n
+0000000363 00000 n
+0000002117 00000 n
+0000002137 00000 n
+trailer
+<<
+/Root 1 0 R
+/Info 3 0 R
+/Size 8
+>>
+startxref
+2235
+%%EOF
