+++ headers/GUI_Call.h
@@ -1,0 +1,53 @@
+#ifndef GUI_CALL_H_INCLUDED
+#define GUI_CALL_H_INCLUDED
+
+void MyButton1_Clicked(GtkButton *b);
+void MyButton2_Clicked(GtkButton *b);
+
+void MyCheckButton1_Toggled(GtkCheckButton *b);
+void MyCheckButton2_Toggled(GtkCheckButton *b);
+void MyCheckButton3_Toggled(GtkCheckButton *b);
+void MyCheckButton4_Toggled(GtkCheckButton *b);
+
+void MyCheckButton5_Toggled(GtkCheckButton *b);
+void MyCheckButton6_Toggled(GtkCheckButton *b);
+void MyCheckButton11_Changed(GtkCheckButton *b);
+void MyEntry2_Changed(GtkCheckButton *b);
+
+void MyRadioButton1_Toggled(GtkRadioButton *b);
+void MyRadioButton2_Toggled(GtkRadioButton *b);
+void MySpinButton1_Changed(GtkSpinButton *s);
+
+void MySpinButton2_Changed(GtkSpinButton *s);
+void MySpinButton8_Changed(GtkSpinButton *s);
+void MySpinButton5_Changed(GtkSpinButton *s);
+void MySpinButton9_Changed(GtkSpinButton *s);
+void MySpinButton3_Changed(GtkSpinButton *s);
+void MySpinButton4_Changed(GtkSpinButton *s);
+void MySpinButton6_Changed(GtkSpinButton *s);
+void MySpinButton7_Changed(GtkSpinButton *s);
+
+void MySpinButton10_Changed(GtkSpinButton *s);
+void MySpinButton11_Changed(GtkSpinButton *s);
+void MySpinButton12_Changed(GtkSpinButton *s);
+void MySpinButton13_Changed(GtkSpinButton *s);
+
+void MyFileChosserButton1_FileSet(GtkFileChooserButton *f);
+void MyEntry1_Changed(GtkEntry *e);
+void MyButton4_Clicked(GtkButton *b);
+
+void MyButton5_Clicked(GtkButton *b);
+
+void MySpinButton14_Changed(GtkSpinButton *b);
+void MySpinButton15_Changed(GtkSpinButton *b);
+void MySpinButton16_Changed(GtkSpinButton *b);
+void MySpinButton17_Changed(GtkSpinButton *b);
+
+void MyCheckButton7_Changed(GtkCheckButton *b);
+void MyCheckButton8_Changed(GtkCheckButton *b);
+void MyCheckButton9_Changed(GtkCheckButton *b);
+void MyCheckButton10_Changed(GtkCheckButton *b);
+
+void MyStackSwitcher2_Changed();
+
+#endif
