+++ headers/GUI_Update.h
@@ -1,0 +1,7 @@
+void WCO_GUI_Update(int i);
+void WCO_GUI_Update_Label(int i);
+void WCO_GUI_Update_Visibility(int i);
+void WCO_GUI_Update_Baseboard();
+void WCO_GUI_Update_DigitRange(int i);
+void WCO_GUI_Update_CreateButton();
+void WCO_GUI_Update_FileName();
