+++ src/main.c
@@ -1,0 +1,42 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <gtk/gtk.h>
+#include <time.h>
+#include <setjmp.h>
+#include <string.h>
+#include <pthread.h>
+#include <unistd.h>
+#include <time.h>
+#include <math.h>
+#include <MagickWand/MagickWand.h>
+
+#include "hpdf.h"
+#include "GUI_Call.h"
+#include "GUI_Status.h"
+#include "GUI_Update.h"
+#include "GUI.h"
+#include "PDF.h"
+#include "PNG.h"
+#include "Worksheet_Baseboard.h"
+#include "Worksheet_Creat_Fraction.h"
+#include "Worksheet_Creat_Task.h"
+
+
+
+int main(void)
+{
+    //Init
+    WCO_PNG_Init();
+    WCO_Worksheet_Baseboard_Init();
+    
+
+    WCO_GUI_Start();
+
+
+
+    return EXIT_SUCCESS;
+}
+
+
+
+
